perf: Use sync instead of copying the directory
**Problem**
We currently need to split the output of the compiler from
the classes directory so we can cache just the output
without mixing it with the resource files
while maintaining similar classes directory as sbt 1.x
(as opposed to just splitting them completely).

However, copying directory is a performance bottleneck.

**Solution**
This uses Sync.sync function to synchronize the backend
directory to classes directory.
This is more reliable than using Analysis information,
since the compilers (and/or plugins) can generate non-class files
on the side in META-INF, or tasty file in case of Scala 3.